local/pyproject.toml
[project]
name = "local"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = []

[dependency-groups]
dev = [
    "aws-cdk-lib==2.186.0",
    "constructs>=10.0.0,<11.0.0",
    "pytest==6.2.5",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
